+++ src/errors/InvalidActionException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Selection\Errors;
+
+class InvalidActionException extends \Exception
+{
+  protected $message = "Erreur : La classe ou la méthode renseignée n'existe pas.";
+}