+++ src/views/components/footer.php
@@ -1,3 +1,7 @@
+<p class="has-text-grey">
+  NOIZET Maxence · 2022
+</p>
+
 </div>
 </div>
 </div>