+++ src/views/components/header.php
@@ -5,11 +5,11 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="icon" type="image/png" href="./src/assets/images/icones8-favicon.png" />
+  <link rel="icon" type="image/png" href="<?= $_ENV["WEB_URL"] . "assets/images/icones8-favicon.png" ?>" />
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@mdi/font@6.5.95/css/materialdesignicons.min.css" />
   <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Questrial&display=swap" />
-  <link rel="stylesheet" href="./src/assets/css/style.css" />
-  <title>[S11] Selection BTS 1.0.0</title>
+  <link rel="stylesheet" href="<?= $_ENV["WEB_URL"] . "assets/css/style.css" ?>" />
+  <title>[S11] Selection BTS <?= $_ENV["APP_VERSION"] ?></title>
 </head>
 
 <body>